+++ R/bootstrap_model-emmeans.R
@@ -1,8 +1,6 @@
 #' @keywords emmeans_methods
 emm_basis.bootstrap_model <- function(object, trms, xlev, grid, ...) {
-  if (!requireNamespace("emmeans", quietly = TRUE)) {
-    stop("this function requires the 'emmeans' package to work.")
-  }
+  insight::check_if_installed("emmeans")
 
   model <- attr(object, "original_model")
   emb <- emmeans::emm_basis(model, trms, xlev, grid, ...)
@@ -18,10 +16,7 @@ emm_basis.bootstrap_model <- function(object, trms, xlev, grid, ...) {
 
 #' @keywords emmeans_methods
 recover_data.bootstrap_model <- function(object, ...) {
-  if (!requireNamespace("emmeans", quietly = TRUE)) {
-    stop("this function requires the 'emmeans' package to work.")
-  }
-
+  insight::check_if_installed("emmeans")
   model <- attr(object, "original_model")
   emmeans::recover_data(model, ...)
 }
@@ -29,20 +24,14 @@ recover_data.bootstrap_model <- function(object, ...) {
 
 #' @keywords emmeans_methods
 emm_basis.bootstrap_parameters <- function(object, trms, xlev, grid, ...) {
-  if (!requireNamespace("emmeans", quietly = TRUE)) {
-    stop("this function requires the 'emmeans' package to work.")
-  }
-
+  insight::check_if_installed("emmeans")
   model <- attr(object, "boot_samples")
   emmeans::emm_basis(model, trms, xlev, grid, ...)
 }
 
 #' @keywords emmeans_methods
 recover_data.bootstrap_parameters <- function(object, ...) {
-  if (!requireNamespace("emmeans", quietly = TRUE)) {
-    stop("this function requires the 'emmeans' package to work.")
-  }
-
+  insight::check_if_installed("emmeans")
   model <- attr(object, "boot_samples")
   emmeans::recover_data(model, ...)
 }