+++ R/extract_parameters.R
@@ -19,8 +19,17 @@
 
   # ==== check if standardization is required and package available
 
+  if (isTRUE(standardize)) {
+    if (verbose) {
+      insight::format_message(warning("'standardize' must be on of 'refit', 'posthoc', 'basic', 'smart' or 'pseudo'.", call. = FALSE))
+    }
+    standardize <- NULL
+  }
+
   if (!is.null(standardize) && !requireNamespace("effectsize", quietly = TRUE)) {
-    insight::print_color("Package 'effectsize' required to calculate standardized coefficients. Please install it.\n", "red")
+    if (verbose) {
+      insight::format_message(warning("Package 'effectsize' required to calculate standardized coefficients. Please install it.", call. = FALSE))
+    }
     standardize <- NULL
   }
 
@@ -81,7 +90,7 @@
 
   if (!is.null(ci)) {
     if (isTRUE(robust)) {
-      ci_df <- suppressMessages(ci_robust(model, ci = ci, verbose = verbose, ...))
+      ci_df <- suppressMessages(ci_robust(model, ci = ci, component = component, verbose = verbose, ...))
     } else if (!is.null(df_method)) {
       ci_df <- suppressMessages(
         ci(
@@ -117,7 +126,7 @@
   # ==== p value
 
   if (isTRUE(robust)) {
-    pval <- p_value_robust(model, ...)
+    pval <- p_value_robust(model, component = component, ...)
   } else if (!is.null(df_method)) {
     pval <- p_value(
       model,
@@ -144,7 +153,7 @@
   std_err <- NULL
 
   if (isTRUE(robust)) {
-    std_err <- standard_error_robust(model, ...)
+    std_err <- standard_error_robust(model, component = component, ...)
   } else if (!is.null(df_method)) {
     std_err <- standard_error(
       model,